% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{get_twilio_credentials}
\alias{get_twilio_credentials}
\title{Get Twilio login credentials.}
\usage{
get_twilio_credentials()
}
\value{
Twilio account SID and auth token
}
\description{
Get Twilio login credentials.
}
\examples{
\dontrun{
  get_twilio_credentials()
}
}

