% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{has_twilio_credentials}
\alias{has_twilio_credentials}
\title{Detects if Twilio login credentials are set}
\usage{
has_twilio_credentials()
}
\value{
TRUE if account sid and auth set, otherwise FALSE
}
\description{
Detects if Twilio login credentials are set
}
\examples{
\dontrun{
  has_twilio_credentials()
}
}

