% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{set_twilio_credentials}
\alias{set_twilio_credentials}
\title{Sets the Twilio login credentials}
\usage{
set_twilio_credentials(account_sid, account_auth)
}
\arguments{
\item{account_sid}{Account SID for Twilio}

\item{account_auth}{Account auth for Twilio}
}
\value{
Twilio account SID and auth token
}
\description{
Sets the Twilio login credentials
}
\examples{
\dontrun{
  set_twilio_credentials("account_sid", "account_auth")
}
}

